% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dates.R
\name{calc_calendar_vars}
\alias{calc_calendar_vars}
\title{Title}
\usage{
calc_calendar_vars(
  .data,
  date_col,
  locale = Sys.getlocale("LC_TIME"),
  week_start = 1
)
}
\arguments{
\item{.data}{}

\item{date_col}{}

\item{locale}{}

\item{week_start}{}
}
\value{
tibble with additional calendar layout variables
}
\description{
Title
}
