% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_events.R
\name{demo_events}
\alias{demo_events}
\title{Generate tibble of demo events}
\usage{
demo_events(n_events = 5, date_range = NULL, max_duration = 5)
}
\arguments{
\item{n_events}{}

\item{date_range}{}

\item{max_duration}{}
}
\value{
tibble
}
\description{
\code{emoji} and \code{details} columns are only generated if the
\code{emoji} and \code{lorem} packages are installed
}
\examples{
demo_events(30)
}
