% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_events_gpt}
\alias{demo_events_gpt}
\title{5 Sample Events generated by ChatGPT 3.5}
\format{
A data frame with 5 (event) rows and 7 columns:
\describe{
\item{event_id}{Unique identifier for the event.}
\item{startDate}{Start date of the event.}
\item{endDate}{End date of the event.}
\item{duration}{Duration of the event.}
\item{event_title}{Title or name of the event.}
\item{event_descr}{Description of the event.}
\item{event_emoji}{Emoji representing the event theme or type.}
\item{event_link}{URL link to the event's website or page.}
}
}
\source{
https://chat.openai.com/share/c68b7a82-5378-45c8-bf41-7fa134f0b74a
}
\usage{
demo_events_gpt
}
\description{
A set of 5 demo events generated by ChatGPT 3.5 on 13 Apr, 2024.
The following field descriptions were also generated in the same session.
}
\keyword{datasets}
