% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_events_overlap}
\alias{demo_events_overlap}
\title{10 Overlapping Events}
\format{
A data frame with 10 (event) rows and 6 columns:
\describe{
\item{event_id}{Unique identifier for the event.}
\item{title}{Plaeholder title for the event}
\item{start}{Start date of the event.}
\item{end}{End date of the event.}
\item{duration}{Duration of the event in days.}
\item{emoji}{Emoji representing the event theme or type.}
\item{details}{Loreum Ipsum placeholder details of the event.}
}
}
\source{
Package Authors
}
\usage{
demo_events_overlap
}
\description{
A set of 10 demo events where Events 6,7,8 and 9 overlap, with placeholder details
generated using the \link[emoji]{zoo} and \link[lorem]{ipsum}.
}
\keyword{datasets}
