% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dates.R
\name{fill_cal_list}
\alias{fill_cal_list}
\title{Fill out event calendar with missing days}
\usage{
fill_cal_list(
  .data,
  date_col,
  cal_range = NULL,
  cal_unit = "day",
  locale = Sys.getlocale("LC_TIME"),
  week_start = 1
)
}
\arguments{
\item{unit}{}
}
\description{
Fill out event calendar with missing days
}
