% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dates.R
\name{fill_missing_units_list}
\alias{fill_missing_units_list}
\title{Fill out event calendar with missing days}
\usage{
fill_missing_units_list(
  .data,
  date_col,
  adjust_months = NULL,
  cal_unit = "day"
)
}
\description{
Fill out event calendar with missing days
}
