% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtilecal-package.R
\docType{package}
\name{ggtilecal-package}
\alias{ggtilecal}
\alias{ggtilecal-package}
\title{ggtilecal: Create static (and interactive) overview calendars using ggplot2}
\description{
Creates tile-based calendar layouts using ggplot2 facets and tiles. Calculates calendar layout positioning variables, and generates customisable ggplot2 calendars for specified time periods and calendar units.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cynthiahqy.github.io/ggtilecal/}
  \item \url{https://github.com/cynthiahqy/ggtilecal}
  \item Report bugs at \url{https://github.com/cynthiahqy/ggtilecal}
}

}
\author{
\strong{Maintainer}: Cynthia A. Huang \email{cynthia.huang@monash.edu} (\href{https://orcid.org/0000-0002-9218-987X}{ORCID})

Authors:
\itemize{
  \item Cynthia A. Huang \email{cynthia.huang@monash.edu} (\href{https://orcid.org/0000-0002-9218-987X}{ORCID})
}

Other contributors:
\itemize{
  \item Christian Klettner [contributor]
}

}
\keyword{internal}
