% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_yearmonth.R
\name{label_yearmonth}
\alias{label_yearmonth}
\title{Label month facets}
\usage{
label_yearmonth(formatstring = "\%b")
}
\arguments{
\item{formatstring}{A character string specifying the format to be used
\itemize{
\item "\%b" is "Jan", "Feb"
\item "\%b \%Y" is "Jan 2024", "Feb 2024"
}}
}
\value{
A labeller function suitable for the \code{labeller} argument in \code{gg_facet_wrap_months()}
}
\description{
This function creates a custom \link[ggplot2]{labeller} for use
in \code{ggplot2::facet_wrap()} to format dates (typically \link[tsibble]{yearmonth}) in a
specified string format.
}
