% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_empty.R
\name{make_empty_month_days}
\alias{make_empty_month_days}
\title{Make padding day unit entries for specified months}
\usage{
make_empty_month_days(
  month_range,
  dates_to = "unit_date",
  adjust_months = c(-0, 0)
)
}
\arguments{
\item{dates_to}{string name for column to output calendar unit dates}

\item{adjust_months}{vector specifying how many months to add}
}
\value{
tibble with one row per day in specified months
}
\description{
Creates empty day units using \code{make_empty_units()} that span the months
encompassing \code{cal_range} with the option to adjust the number of months
either side of \code{cal_range}
}
\examples{
make_empty_month_days(c("2024-03-05", "2024-04-15"))
}
