% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dates.R
\name{make_empty_units}
\alias{make_empty_units}
\title{Make empty sequence of calendar units between specified date range}
\usage{
make_empty_units(cal_range, dates_to = "unit_date", cal_unit = "day")
}
\arguments{
\item{cal_range}{\code{c(start,end)} range for which to create sequence of calendar units}

\item{dates_to}{string name for column to output calendar unit dates}

\item{cal_unit}{increment of calendar sequence:
\verb{"day", "week", "month", "quarter", "year"}. See \code{\link[base]{seq.Date}}}
}
\value{
tibble
}
\description{
Make empty sequence of calendar units between specified date range
}
\examples{
make_empty_units(c("2024-03-05", "2024-04-15"))
}
