% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_facet_wrap_months.R
\name{theme_bw_tilecal}
\alias{theme_bw_tilecal}
\title{A modified version of \code{ggplot2::theme_bw()} for calendar layouts}
\usage{
theme_bw_tilecal(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
A modified version of \code{ggplot2::theme_bw()} for calendar layouts
}
