% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_calendar_vars.R
\name{calc_calendar_vars}
\alias{calc_calendar_vars}
\title{Calculate variables for calendar layout}
\usage{
calc_calendar_vars(
  .data,
  date_col,
  locale = Sys.getlocale("LC_TIME"),
  week_start = 1
)
}
\arguments{
\item{.data}{tibble or data frame containing dates to be plotted in calendar layout}

\item{date_col}{column containing calendar unit dates}

\item{locale}{locale to use for day names. Default to current locale.}

\item{week_start}{day on which week starts following ISO conventions: 1 means Monday
and 7 means Sunday (default). When \code{label = FALSE} and \code{week_start = 7}, the number
returned for Sunday is 1, for Monday is 2, etc. When \code{label = TRUE}, the returned
value is a factor with the first level being the week start (e.g. Sunday if
\code{week_start = 7}). You can set \code{lubridate.week.start} option to control this
parameter globally.}
}
\value{
tibble with additional calendar layout variables
}
\description{
Helper function for calculating calendar layout variables.
}
\examples{
make_empty_month_days(c("2024-01-01", "2024-02-01")) |> 
  calc_calendar_vars(unit_date)
}
