% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dates.R
\name{fill_missing_units}
\alias{fill_missing_units}
\title{Fill out event calendar with missing days}
\usage{
fill_missing_units(.data, date_col, cal_range = NULL, cal_unit = "day")
}
\arguments{
\item{unit}{}
}
\value{
tibble
}
\description{
Fill out event calendar with missing days
}
