% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dates.R
\name{fill_missing_units}
\alias{fill_missing_units}
\title{Fill out long format event table with missing dates}
\usage{
fill_missing_units(
  .events_long,
  date_col,
  adjust_months = NULL,
  cal_unit = "day"
)
}
\arguments{
\item{.events_long}{long format calendar event data}

\item{date_col}{column containing calendar unit dates}

\item{adjust_months}{how many months to add before and after}

\item{cal_unit}{increment of calendar sequence passed to \code{by} argument in
\code{\link[base]{seq.Date}}}
}
\value{
tibble
}
\description{
Helper function for filling out event table with any missing calendar units.
}
\examples{
demo_events_gpt |>
  reframe_events(startDate, endDate) |>
  fill_missing_units(unit_date)
}
