% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{gg_facet_wrap_months}
\alias{gg_facet_wrap_months}
\title{Make Monthly Calendar Facets}
\usage{
gg_facet_wrap_months(
  .events_long,
  date_col,
  cal_range = NULL,
  unit = "day",
  locale = NULL,
  week_start = NULL,
  nrow = NULL,
  ncol = NULL,
  .geom = list(geom_tile(), geom_text(nudge_y = 0.25)),
  .scale_coord = list(scale_y_reverse(), scale_x_discrete(position = "top"),
    coord_fixed(expand = TRUE)),
  .theme = list(theme_ggtilecal()),
  .layers = list()
)
}
\arguments{
\item{cal_range}{WIP/TODO}

\item{unit}{TODO rename to \code{tile_unit}}

\item{locale}{locale to use for day names. Default to current locale.}

\item{week_start}{day on which week starts following ISO conventions: 1 means Monday
and 7 means Sunday (default). When \code{label = FALSE} and \code{week_start = 7}, the number
returned for Sunday is 1, for Monday is 2, etc. When \code{label = TRUE}, the returned
value is a factor with the first level being the week start (e.g. Sunday if
\code{week_start = 7}). You can set \code{lubridate.week.start} option to control this
parameter globally.}

\item{nrow, ncol}{Number of rows and columns.}

\item{.geom, .scale_coord, .theme, .layers}{Customisable lists of ggplot2 layers to add to the plot.
An empty list() leaves the plot unmodified.}
}
\value{
ggplot
}
\description{
Generates calendar with monthly facets by:
\itemize{
\item Padding event list with any missing days
\item Calculating variables for calendar layout
\item Returning a ggplot object as per Details.
}
}
\details{
Returns a ggplot with the following fixed layers specified (in order):
\itemize{
\item \code{aes()} mapping using calculated vars:
\itemize{
\item \code{x} is day of week, \code{y} is week in month, \code{label} is day of month
}
\item \code{facet_wrap()} by month
\item \code{labs()} to remove calculated vars axis labels
}

and default customisable layers:
\itemize{
\item \code{geom_tile()}, \code{geom_text()} to label each day
\item \code{scale_y_reverse()} to order day in month correctly
\item \code{scale_x_discrete()} to position weekday labels
\item \code{coord_fixed()} to square each tile
\item \code{theme_ggtilecal()} to apply sensible theme defaults
}

To modify layers use \code{.geom}, \code{.scale_coord}, \code{.theme} or \code{.layers}.
This can be used to add interactive geoms (e.g. from \code{ggiraph})
}
\examples{
library(dplyr)
library(ggplot2)
demo_events(20) |>
  reframe_intervals(start, end) |>
  group_by(unit_date) |>
  slice_min(order_by = duration) |>
  gg_facet_wrap_months(unit_date)
}
