% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_facet_wrap_months.R
\name{gg_facet_wrap_months}
\alias{gg_facet_wrap_months}
\title{Make Monthly Calendar Facets}
\usage{
gg_facet_wrap_months(
  .events_long,
  date_col,
  locale = NULL,
  week_start = NULL,
  nrow = NULL,
  ncol = NULL,
  .geom = list(geom_tile(color = "grey70", fill = "transparent"), geom_text(nudge_y =
    0.25)),
  .scale_coord = list(scale_y_reverse(), scale_x_discrete(position = "top"),
    coord_fixed(expand = TRUE)),
  .theme = list(theme_bw_tilecal()),
  .other = list()
)
}
\arguments{
\item{.events_long}{long format calendar event data}

\item{date_col}{column containing calendar unit dates}

\item{locale}{locale to use for day names. Default to current locale.}

\item{week_start}{day on which week starts following ISO conventions: 1 means Monday
and 7 means Sunday (default). When \code{label = FALSE} and \code{week_start = 7}, the number
returned for Sunday is 1, for Monday is 2, etc. When \code{label = TRUE}, the returned
value is a factor with the first level being the week start (e.g. Sunday if
\code{week_start = 7}). You can set \code{lubridate.week.start} option to control this
parameter globally.}

\item{nrow, ncol}{Number of rows and columns.}

\item{.geom, .scale_coord, .theme, .other}{Customisable lists of ggplot2 components to add to the plot.
An empty \code{list()} leaves the plot unmodified.}
}
\value{
ggplot
}
\description{
Generates calendar with monthly facets by:
\itemize{
\item Padding event list with any missing days via \code{fill_missing_units()}
\item Calculating variables for calendar layout via \code{calc_calendar_vars()}
\item Returning a ggplot object as per Details.
}
}
\details{
Returns a ggplot with the following fixed components
using calculated layout variables:
\itemize{
\item \code{aes()} mapping:
\itemize{
\item \code{x} is day of week,
\item \code{y} is week in month,
\item \code{label} is day of month
}
\item \code{facet_wrap()} by month
\item \code{labs()} to remove axis labels for calculated layout variables
}

and default customisable components:
\itemize{
\item \code{geom_tile()}, \code{geom_text()} to label each day which inherit calculated variables
\item \code{scale_y_reverse()} to order day in month correctly
\item \code{scale_x_discrete()} to position weekday labels
\item \code{coord_fixed()} to square each tile
\item \code{theme_bw_tilecal()} to apply sensible theme defaults
}

To modify components alter the \code{.geom} and \code{.scale_coord},
which inherit the calculate layout mapping by default
(via the ggplot2 \code{inherit.aes} argument).

To additional components use the ggplot \code{+} function as normal,
or pass components to the \code{.other} argument.
This can be used to add interactive geoms (e.g. from \code{ggiraph})

To modify the theme, use the ggplot \code{+} function as normal,
or add additional elements to the list in \code{.theme}.

To remove any of the optional components, set the argument to any empty \code{list()}
}
\examples{
library(dplyr)
library(ggplot2)

demo_events_gpt |>
  reframe_events(startDate, endDate) |>
  group_by(unit_date) |>
  slice_min(order_by = duration) |>
  gg_facet_wrap_months(unit_date) +
  geom_text(aes(label = event_emoji), nudge_y = -0.25, na.rm = TRUE)
}
