% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtilecal-package.R
\docType{package}
\name{ggtilecal-package}
\alias{ggtilecal}
\alias{ggtilecal-package}
\title{ggtilecal: Package for creating interactive and static overview calendars using tiles}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cynthiahqy.github.io/ggtilecal/}
}

}
\author{
\strong{Maintainer}: Cynthia Huang \email{cynthia.huang@monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
