% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reframe_events.R
\name{reframe_events}
\alias{reframe_events}
\title{Reframe event intervals to calendar units}
\usage{
reframe_events(
  .data,
  event_start,
  event_end,
  dates_to = "unit_date",
  cal_unit = "day"
)
}
\arguments{
\item{.data}{A data frame or tibble containing event details}

\item{event_start}{column containing event start date}

\item{event_end}{column containing event end date}

\item{dates_to}{string name for column to output calendar unit dates}

\item{cal_unit}{increment of calendar sequence passed to \code{by} argument in \code{\link[base]{seq.Date}}}
}
\value{
See return value of \link[dplyr]{reframe}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Convert rows of events to sequence of calendar units for plotting.
Duplicates all columns except for \code{event_start} and \code{event_date}.

This is a thin wrapper around \code{dplyr::reframe()}.
}
\examples{
demo_events(5) |> 
  reframe_events(start, end)
}
